ref(releases): Move releases endpoints and tests to releases folder
Moves all release-related endpoints from src/sentry/api/endpoints/
to src/sentry/releases/endpoints/ and corresponding test files from
tests/sentry/api/endpoints/ to tests/sentry/releases/endpoints/.

Updates all import references across the codebase to use the new paths.

refs https://linear.app/getsentry/issue/RTC-1110/move-issues-endpoints-into-issues-app
