+++ src/sentry/api/helpers/source_map_helper.py
@@ -7,7 +7,6 @@ from packaging.version import Version
 from rest_framework.exceptions import NotFound, ParseError
 
 from sentry import eventstore
-from sentry.api.endpoints.project_release_files import ArtifactSource
 from sentry.debug_files.release_files import maybe_renew_releasefiles
 from sentry.eventstore.models import BaseEvent
 from sentry.interfaces.exception import Exception as ExceptionInterface
@@ -17,6 +16,7 @@ from sentry.models.project import Project
 from sentry.models.release import Release
 from sentry.models.releasefile import ReleaseFile, read_artifact_index
 from sentry.models.sourcemapprocessingissue import SourceMapProcessingIssue
+from sentry.releases.endpoints.project_release_files import ArtifactSource
 from sentry.utils.javascript import find_sourcemap
 from sentry.utils.urls import non_standard_url_join
