+++ src/sentry/releases/endpoints/organization_release_details.py
@@ -11,11 +11,6 @@ from sentry.api.api_owners import ApiOwner
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import ReleaseAnalyticsMixin, region_silo_endpoint
 from sentry.api.bases.organization import OrganizationReleasesBaseEndpoint
-from sentry.api.endpoints.organization_releases import (
-    _release_suffix,
-    add_environment_to_queryset,
-    get_stats_period_detail,
-)
 from sentry.api.exceptions import ConflictError, InvalidRepository, ResourceDoesNotExist
 from sentry.api.serializers import serialize
 from sentry.api.serializers.rest_framework import (
@@ -38,6 +33,11 @@ from sentry.models.organization import Organization
 from sentry.models.project import Project
 from sentry.models.release import Release, ReleaseStatus
 from sentry.models.releases.exceptions import ReleaseCommitError, UnsafeReleaseDeletion
+from sentry.releases.endpoints.organization_releases import (
+    _release_suffix,
+    add_environment_to_queryset,
+    get_stats_period_detail,
+)
 from sentry.snuba.sessions import STATS_PERIODS
 from sentry.types.activity import ActivityType
 from sentry.utils.sdk import bind_organization_context
