+++ src/sentry/releases/endpoints/organization_release_file_details.py
@@ -5,10 +5,10 @@ from rest_framework.response import Response
 from sentry.api.api_publish_status import ApiPublishStatus
 from sentry.api.base import region_silo_endpoint
 from sentry.api.bases.organization import OrganizationReleasesBaseEndpoint
-from sentry.api.endpoints.project_release_file_details import ReleaseFileDetailsMixin
 from sentry.api.exceptions import ResourceDoesNotExist
 from sentry.models.organization import Organization
 from sentry.models.release import Release
+from sentry.releases.endpoints.project_release_file_details import ReleaseFileDetailsMixin
 
 
 class ReleaseFileSerializer(serializers.Serializer):
