+++ tests/sentry/api/serializers/test_release.py
@@ -8,7 +8,6 @@ from uuid import uuid4
 from rest_framework.exceptions import ErrorDetail
 
 from sentry import tagstore
-from sentry.api.endpoints.organization_releases import ReleaseSerializerWithProjects
 from sentry.api.serializers import serialize
 from sentry.api.serializers.models.release import GroupEventReleaseSerializer, get_users_for_authors
 from sentry.models.commit import Commit
@@ -19,6 +18,7 @@ from sentry.models.release import Release
 from sentry.models.releasecommit import ReleaseCommit
 from sentry.models.releaseprojectenvironment import ReleaseProjectEnvironment, ReleaseStages
 from sentry.models.releases.release_project import ReleaseProject
+from sentry.releases.endpoints.organization_releases import ReleaseSerializerWithProjects
 from sentry.silo.base import SiloMode
 from sentry.testutils.cases import SnubaTestCase, TestCase
 from sentry.testutils.helpers.datetime import before_now
