+++ tests/sentry/releases/endpoints/test_project_release_details.py
