+++ tests/sentry/releases/endpoints/test_project_release_stats.py
